+++ Sources/PeripheryKit/ScanResult.swift
@@ -4,7 +4,7 @@ public struct ScanResult {
     enum Annotation {
         case unused
         case assignOnlyProperty
-        case redundantProtocol(references: Set<Reference>)
+        case redundantProtocol(references: Set<Reference>, inherited: Set<String>)
         case redundantPublicAccessibility(modules: Set<String>)
     }