+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - The option `--external-encodable-protocols` is deprecated, use `--external-codable-protocols` instead.
 - Assign-only properties on structs with synthesized initializers are now detected.
 - Added the `--retain-codable-properties` option to retain all properties on Codable types.
+- Results for redundant protocol conformances will now list the inherited protocols that should replace the redundant conformance.
 
 ##### Bug Fixes